Allen's changes with the database
