--- Swift/FishSense/Local Database/Info.plist
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-    <key>UIApplicationSceneManifest</key>
-    <dict>
-        <key>UIApplicationSupportsMultipleScenes</key>
-        <false/>
-        <key>UISceneConfigurations</key>
-        <dict>
-            <key>UIWindowSceneSessionRoleApplication</key>
-            <array>
-                <dict>
-                    <key>UISceneConfigurationName</key>
-                    <string>Default Configuration</string>
-                    <key>UISceneDelegateClassName</key>
-                    <string>$(PRODUCT_MODULE_NAME).SceneDelegate</string>
-                    <key>UISceneStoryboardFile</key>
-                    <string>Main</string>
-                </dict>
-            </array>
-        </dict>
-    </dict>
-    <!-- Disable state restoration -->
-    <key>UIApplicationStateRestorationEnabled</key>
-    <false/>
-    <!-- Permissions -->
-    <key>NSDocumentsFolderUsageDescription</key>
-    <string>This app needs access to the Documents folder to store data.</string>
-    <key>NSFileProviderDomainUsageDescription</key>
-    <string>This app needs access to the FileProvider domain to manage files.</string>
-</dict>
-</plist>
